requirements.txt
fastapi
uvicorn[standard]
praw
vaderSentiment
sqlalchemy
pydantic
python-dotenv
apscheduler